Adjust test execution error handling.
